requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
scikit-learn>=1.4.0
numpy>=1.24.0
pandas>=2.0.0

